refactor vote inbound outbound into separate files
